Crate webbrowser

Open URLs in the web browsers available on a platform

Inspired by the webbrowser python library

Examples

use webbrowser;
 
if webbrowser::open("http://github.com").is_ok() {
    // ...
}

Currently state of platform support is:

macos => default, as well as browsers listed under Browser

windows => default browser only

linux => default browser only

android => not supported right now

ios => not supported right now

Important note:

This library requires availability of browsers and a graphical environment during runtime

cargo test will actually open the browser locally

Enums

Browser
Browser types available

Functions

open
Opens the URL on the default browser of this platform

open_browser
Opens the specified URL on the specific browser (if available) requested. Return semantics are the same as for open